add action on Share option on More screen
