+++ app/src/main/kotlin/com/bunbeauty/tiptoplive/features/more/presentation/More.kt
@@ -18,10 +18,12 @@ interface More {
 
     sealed interface Action: Base.Action {
         data object PremiumClick: Action
+        data object ShareClick: Action
     }
 
     sealed interface Event: Base.Event {
         data object NavigateToPremiumDetails: Event
+        data object OpenSharing: Event
     }
